+++ tests/testthat/test.predicted_draws.R
@@ -16,7 +16,7 @@ context("predicted_draws")
 # data
 mtcars_tbl = mtcars %>%
   set_rownames(seq_len(nrow(.))) %>%
-  as_data_frame()
+  as_tibble()
 
 
 test_that("[add_]predicted_draws throws an error on unsupported models", {
@@ -41,7 +41,7 @@ test_that("[add_]predicted_draws and basic arguments works on a simple rstanarm 
       .draw = seq_len(n())
     ) %>%
     gather(.row, .prediction, -.chain, -.iteration, -.draw) %>%
-    as_data_frame()
+    as_tibble()
 
   ref = mtcars_tbl %>%
     mutate(.row = rownames(.)) %>%
@@ -65,7 +65,7 @@ test_that("[add_]predicted_draws and basic arguments works on an rstanarm model 
       .draw = seq_len(n())
     ) %>%
     gather(.row, .prediction, -.chain, -.iteration, -.draw) %>%
-    as_data_frame()
+    as_tibble()
 
   ref = mtcars_tbl %>%
     mutate(.row = rownames(.)) %>%
@@ -91,7 +91,7 @@ test_that("[add_]predicted_draws works on a simple brms model", {
       .draw = seq_len(n())
     ) %>%
     gather(.row, .prediction, -.chain, -.iteration, -.draw) %>%
-    as_data_frame()
+    as_tibble()
 
   ref = mtcars_tbl %>%
     mutate(.row = rownames(.)) %>%
@@ -100,6 +100,31 @@ test_that("[add_]predicted_draws works on a simple brms model", {
 
   expect_equal(predicted_draws(m_hp, mtcars_tbl, n = 100, seed = 123), ref)
   expect_equal(add_predicted_draws(mtcars_tbl, m_hp, n = 100, seed = 123), ref)
+})
+
+test_that("[add_]predicted_draws works on brms models with categorical outcomes", {
+  skip_if_not_installed("brms")
+  m_cyl_mpg = readRDS("../models/models.brms.m_cyl_mpg.rds")
+
+  set.seed(1234)
+  preds = predict(m_cyl_mpg, mtcars_tbl, summary = FALSE, nsamples = 100) %>%
+    array2df(list(.draw = NA, .row = NA), label.x = ".prediction") %>%
+    mutate(
+      .chain = NA_integer_,
+      .iteration = NA_integer_,
+      .draw = as.integer(.draw),
+      .row = as.character(.row),
+      .prediction = factor(.prediction, levels = 1:3, labels = paste0("c", c(4,6,8)))
+    )
+
+  ref = mtcars_tbl %>%
+    mutate(.row = rownames(.)) %>%
+    inner_join(preds, by = ".row") %>%
+    mutate(.row = as.integer(.row)) %>%
+    group_by(mpg, cyl, disp, hp, drat, wt, qsec, vs, am, gear, carb, .row)
+
+  expect_equal(predicted_draws(m_cyl_mpg, mtcars_tbl, seed = 1234, n = 100), ref)
+  expect_equal(add_predicted_draws(mtcars_tbl, m_cyl_mpg, seed = 1234, n = 100), ref)
 })
 
 test_that("[add_]predicted_draws throws an error when nsamples is called instead of n in brms", {